feat(donations): implement donation management dialog
- replaces add-donation-dialog with a more flexible donation-dialog component
- integrates donation-dialog with data table for adding and editing donations
- enhances data-table-actions-menu to support donation actions
- removes redundant code and improves ui/ux for donation management
