+++ app/(authenticated)/dashboard/donations/columns.tsx
@@ -1,51 +1,18 @@
 'use client';
 
 import { ColumnDef } from '@tanstack/react-table';
-import { ArrowUpDown, Edit, MoreHorizontal } from 'lucide-react';
+import { ArrowUpDown } from 'lucide-react';
 import { Button } from '@/components/ui/button';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from '@/components/ui/dialog';
-import React from 'react';
 import { Badge } from '@/components/ui/badge';
 import { DonationColumn } from '@/types/donations';
+import { DataTableActionsMenu } from '@/components/data-table/actions-menu';
+import { DonationDialog } from '@/components/dialogs/donation-dialog';
 
 export const columns: ColumnDef<DonationColumn>[] = [
   {
     id: 'edit',
     cell: ({ row }) => {
-      const [open, setOpen] = React.useState(false);
-
-      return (
-        <Dialog open={open} onOpenChange={setOpen}>
-          <DialogTrigger asChild>
-            <div className="cursor-pointer hover:underline">
-              <Edit className="h-4 w-4" />
-            </div>
-          </DialogTrigger>
-          <DialogContent>
-            <DialogHeader>
-              <DialogTitle>Donor Details</DialogTitle>
-              <DialogDescription>
-                Enter the donor details below
-              </DialogDescription>
-            </DialogHeader>
-          </DialogContent>
-        </Dialog>
-      );
+      return <DonationDialog donationData={row} />;
     },
     enableHiding: false,
   },
@@ -152,23 +119,7 @@ export const columns: ColumnDef<DonationColumn>[] = [
   {
     id: 'actions',
     cell: ({ row }) => {
-      const payment = row.original;
-
-      return (
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant="ghost" className="h-8 w-8 p-0">
-              <span className="sr-only">Open menu</span>
-              <MoreHorizontal className="h-4 w-4" />
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align="end">
-            <DropdownMenuLabel className="pb-2">Actions</DropdownMenuLabel>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem>Delete</DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      );
+      return <DataTableActionsMenu row={row} />;
     },
   },
 ];
